# docs/spec/api/openapi.yaml
openapi: 3.1.0
info:
  title: IEapp API
  version: "0.1"
  description: |
    Minimal OpenAPI document for the current REST surface.
servers:
  - url: http://localhost:8000
paths:
...
  /api/workspaces/{workspace_id}/notes:
    get:
      summary: List notes
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Create note
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /api/workspaces/{workspace_id}/notes/{note_id}:
    get:
      summary: Get note
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update note
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "409":
          description: Conflict
    delete:
      summary: Delete note
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: note_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/workspaces/{workspace_id}/query:
    post:
      summary: Query notes
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/workspaces/{workspace_id}/search:
    get:
      summary: Keyword search
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...